:truck: Moved LinkAnchor utility to separate file
