+++ @navikt/core/react/src/link-card/LinkCard.tsx
@@ -1,27 +1,18 @@
-/* eslint-disable jsx-a11y/no-static-element-interactions */
-
-/* eslint-disable jsx-a11y/click-events-have-key-events */
 import React, {
-  AnchorHTMLAttributes,
   HTMLAttributes,
   ImgHTMLAttributes,
   SVGProps,
   forwardRef,
-  useCallback,
-  useRef,
 } from "react";
 import { ArrowRightIcon } from "@navikt/aksel-icons";
-import { Slot } from "../slot/Slot";
 import { useRenameCSS } from "../theme/Theme";
 import { BodyLong, Heading } from "../typography";
-import { createContext } from "../util/create-context";
-import { useMergeRefs } from "../util/hooks";
-
-/**
- * TODO:
- * - Consider the need for `as` with overridablecomponent-API vs Slot, or both.
- * Especially for the `LinkCardAnchor` component where most relevant to use other frameworks for links.
- */
+import {
+  LinkAnchor,
+  LinkAnchorOverlay,
+  type LinkAnchorOverlayProps,
+  LinkAnchorProps,
+} from "./LinkAnchor";
 
 /* ------------------------------ LinkCard Root ----------------------------- */
 type LinkCardProps = LinkAnchorOverlayProps & {
@@ -241,104 +232,6 @@ const LinkCardImage = forwardRef<HTMLImageElement, LinkCardImageProps>(
   },
 );
 
-/* ------------------------ LinkCard Anchor utilities ----------------------- */
-type LinkAnchorOverlayContextProps = {
-  anchorRef: React.RefObject<HTMLAnchorElement>;
-};
-
-const [LinkOverlayContextProvider, useLinkOverlayContext] =
-  createContext<LinkAnchorOverlayContextProps>({
-    name: "LinkAnchorOverlayContext",
-  });
-
-interface LinkAnchorOverlayProps extends HTMLAttributes<HTMLDivElement> {
-  children: React.ReactNode;
-  asChild?: boolean;
-}
-
-const LinkAnchorOverlay = forwardRef<HTMLDivElement, LinkAnchorOverlayProps>(
-  (
-    { children, asChild, className, ...restProps }: LinkAnchorOverlayProps,
-    forwardedRef,
-  ) => {
-    const { cn } = useRenameCSS();
-    const anchorRef = useRef<HTMLAnchorElement>(null);
-
-    const isTextSelected = useCallback(() => {
-      return !!window.getSelection()?.toString();
-    }, []);
-
-    const Component = asChild ? Slot : "div";
-
-    return (
-      <LinkOverlayContextProvider anchorRef={anchorRef}>
-        <Component
-          ref={forwardedRef}
-          {...restProps}
-          className={cn("navds-link-anchor__overlay", className)}
-          onClick={(e) => {
-            if (e.target === anchorRef.current || isTextSelected()) {
-              return;
-            }
-
-            const event = new MouseEvent("click", {
-              bubbles: true,
-              cancelable: true,
-              view: window,
-              ctrlKey: e.ctrlKey,
-              shiftKey: e.shiftKey,
-              altKey: e.altKey,
-              metaKey: e.metaKey,
-              button: e.button,
-              screenX: e.screenX,
-              screenY: e.screenY,
-              clientX: e.clientX,
-              clientY: e.clientY,
-            });
-
-            anchorRef.current?.dispatchEvent(event);
-          }}
-        >
-          {children}
-        </Component>
-      </LinkOverlayContextProvider>
-    );
-  },
-);
-
-interface LinkAnchorProps extends AnchorHTMLAttributes<HTMLAnchorElement> {
-  children: React.ReactNode;
-  asChild?: boolean;
-}
-
-const LinkAnchor = forwardRef<HTMLAnchorElement, LinkAnchorProps>(
-  (
-    { children, asChild, className, ...restProps }: LinkAnchorProps,
-    forwardedRef,
-  ) => {
-    const { cn } = useRenameCSS();
-
-    const context = useLinkOverlayContext(false);
-    const mergedRefs = useMergeRefs(forwardedRef, context?.anchorRef);
-
-    const Component = asChild ? Slot : "a";
-
-    return (
-      <Component
-        ref={mergedRefs}
-        {...restProps}
-        className={cn("navds-link-anchor", className)}
-        onClick={(e) => {
-          console.info("LinkAnchor clicked", e);
-          /* e.preventDefault(); */
-        }}
-      >
-        {children}
-      </Component>
-    );
-  },
-);
-
 /* --------------------------- LinkCard utilities --------------------------- */
 function aspectRatioClassName(aspectRatio?: ImageAspectRatio): string {
   if (!aspectRatio) {
@@ -359,8 +252,6 @@ export {
   LinkCardIcon,
   LinkCardImage,
   LinkCardTitle,
-  LinkAnchor,
-  LinkAnchorOverlay,
 };
 
 export type {
